+++ CotEditor/Sources/Document Window/Text View/EditorTextView.swift
@@ -99,6 +99,7 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, MultiCursorEdit
     
     var commentDelimiters: Syntax.Comment = Syntax.Comment()
     var commentsOutAfterIndent: Bool = false
+    var commentSpacer: String = ""
     var syntaxCompletionWords: [String] = []
     var completionWordTypes: CompletionWordTypes = []