+++ Packages/EditorCore/Sources/TextEditing/String+Commenting.swift
@@ -65,18 +65,19 @@ public extension String {
     /// - Parameters:
     ///   - types: The commenting style to apply. If `.both` is specified, inline comments take precedence over block comments.
     ///   - delimiters: The comment delimiters to apply.
+    ///   - spacer: The spacer between delimiter and string.
     ///   - selectedRanges: The current selected ranges in the editor.
     ///   - location: The location type to insert comment delimiters.
-    func commentOut(types: CommentTypes, delimiters: Syntax.Comment, in selectedRanges: [NSRange], at location: CommentOutLocation) -> EditingContext? {
+    func commentOut(types: CommentTypes, delimiters: Syntax.Comment, spacer: String, in selectedRanges: [NSRange], at location: CommentOutLocation) -> EditingContext? {
         
         guard !delimiters.isEmpty else { return nil }
         
         let items: [NSRange.InsertionItem] = {
             if types.contains(.inline), let delimiter = delimiters.inline {
-                return self.inlineCommentOut(delimiter: delimiter, ranges: selectedRanges, at: location)
+                return self.inlineCommentOut(delimiter: delimiter, spacer: spacer, ranges: selectedRanges, at: location)
             }
             if types.contains(.block), let delimiters = delimiters.block {
-                return self.blockCommentOut(delimiters: delimiters, ranges: selectedRanges, at: location)
+                return self.blockCommentOut(delimiters: delimiters, spacer: spacer, ranges: selectedRanges, at: location)
             }
             return []
         }()
@@ -96,20 +97,21 @@ public extension String {
     /// - Parameters:
     ///   - delimiters: The comment delimiters to remove.
     ///   - selectedRanges: The current selected ranges in the editor.
-    func uncomment(delimiters: Syntax.Comment, in selectedRanges: [NSRange]) -> EditingContext? {
+    func uncomment(delimiters: Syntax.Comment, spacer: String, in selectedRanges: [NSRange]) -> EditingContext? {
         
         guard !delimiters.isEmpty else { return nil }
         
         let deletionRanges: [NSRange] = {
+            
             if let delimiters = delimiters.block {
                 let targetRanges = selectedRanges.map { $0.isEmpty ? self.lineContentsRange(for: $0) : $0 }.uniqued
-                if let ranges = self.rangesOfBlockDelimiters(delimiters, ranges: targetRanges) {
+                if let ranges = self.rangesOfBlockDelimiters(delimiters, spacer: spacer, ranges: targetRanges) {
                     return ranges
                 }
             }
             if let delimiter = delimiters.inline {
-                let targetRanges = selectedRanges.map(self.lineContentsRange(for:)).uniqued
-                if let ranges = self.rangesOfInlineDelimiter(delimiter, ranges: targetRanges) {
+                let targetRanges = selectedRanges.map (self.lineContentsRange(for:)).uniqued
+                if let ranges = self.rangesOfInlineDelimiter(delimiter, spacer: spacer, ranges: targetRanges) {
                     return ranges
                 }
             }
@@ -145,13 +147,13 @@ public extension String {
         guard !targetRanges.isEmpty else { return false }
         
         if let delimiters = delimiters.block,
-           let ranges = self.rangesOfBlockDelimiters(delimiters, ranges: targetRanges)
+           let ranges = self.rangesOfBlockDelimiters(delimiters, spacer: "", ranges: targetRanges)
         {
             return partly ? true : (ranges.count == (2 * targetRanges.count))
         }
         
         if let delimiter = delimiters.inline,
-           let ranges = self.rangesOfInlineDelimiter(delimiter, ranges: targetRanges)
+           let ranges = self.rangesOfInlineDelimiter(delimiter, spacer: "", ranges: targetRanges)
         {
             let lineRanges = targetRanges.flatMap(self.lineContentsRanges(for:)).uniqued
             return partly ? true : (ranges.count == lineRanges.count)
@@ -168,10 +170,11 @@ extension String {
     ///
     /// - Parameters:
     ///   - delimiter: The inline comment delimiter to insert.
+    ///   - spacer: The spacer between delimiter and string.
     ///   - ranges: The ranges whose lines should be commented out.
     ///   - location: The location type to insert comment delimiters.
     /// - Returns: Items that contain editing information to insert comment delimiters.
-    func inlineCommentOut(delimiter: String, ranges: [NSRange], at location: CommentOutLocation) -> [NSRange.InsertionItem] {
+    func inlineCommentOut(delimiter: String, spacer: String = "", ranges: [NSRange], at location: CommentOutLocation) -> [NSRange.InsertionItem] {
         
         let locations = switch location {
             case .selection:
@@ -188,7 +191,7 @@ extension String {
                 self.minimumCommonIndentationLocations(for: ranges, tabWidth: tabWidth)
         }
         
-        return locations.map { NSRange.InsertionItem(string: delimiter, location: $0, forward: true) }
+        return locations.map { NSRange.InsertionItem(string: delimiter + spacer, location: $0, forward: true) }
     }
     
     
@@ -196,10 +199,11 @@ extension String {
     ///
     /// - Parameters:
     ///   - delimiters: The pair of block comment delimiters to insert.
+    ///   - spacer: The spacer between delimiter and string.
     ///   - ranges: The ranges where to comment out.
     ///   - location: The location type to insert comment delimiters.
     /// - Returns: Items that contain editing information to insert comment delimiters.
-    func blockCommentOut(delimiters: Pair<String>, ranges: [NSRange], at location: CommentOutLocation) -> [NSRange.InsertionItem] {
+    func blockCommentOut(delimiters: Pair<String>, spacer: String = "", ranges: [NSRange], at location: CommentOutLocation) -> [NSRange.InsertionItem] {
         
         let targetRanges = switch location {
             case .selection:
@@ -219,8 +223,8 @@ extension String {
         }
         
         return targetRanges.flatMap {
-            [NSRange.InsertionItem(string: delimiters.begin, location: $0.lowerBound, forward: true),
-             NSRange.InsertionItem(string: delimiters.end, location: $0.upperBound, forward: false)]
+            [NSRange.InsertionItem(string: delimiters.begin + spacer, location: $0.lowerBound, forward: true),
+             NSRange.InsertionItem(string: spacer + delimiters.end, location: $0.upperBound, forward: false)]
         }
     }
     
@@ -229,16 +233,18 @@ extension String {
     ///
     /// - Parameters:
     ///   - delimiter: The inline delimiter to find.
+    ///   - spacer: The spacer between delimiter and string.
     ///   - ranges: The ranges where to find.
-    /// - Returns: Ranges where delimiters are, or `nil` when no delimiters was found.
-    func rangesOfInlineDelimiter(_ delimiter: String, ranges: [NSRange]) -> [NSRange]? {
+    /// - Returns: Ranges where delimiters and spacers are, or `nil` when no delimiters was found.
+    func rangesOfInlineDelimiter(_ delimiter: String, spacer: String, ranges: [NSRange]) -> [NSRange]? {
         
         let ranges = ranges.filter { !$0.isEmpty }
         
         guard !ranges.isEmpty, !self.isEmpty else { return [] }
         
         let delimiterPattern = NSRegularExpression.escapedPattern(for: delimiter)
-        let pattern = "^[ \t]*(\(delimiterPattern))"
+        let spacerPattern = spacer.isEmpty ? "" : "(?:\(spacer))?"
+        let pattern = "^[ \t]*(\(delimiterPattern + spacerPattern))"
         let regex = try! NSRegularExpression(pattern: pattern, options: [.anchorsMatchLines])
         
         let delimiterRanges = ranges
@@ -256,9 +262,10 @@ extension String {
     ///
     /// - Parameters:
     ///   - delimiters: The pair of block delimiters to find.
+    ///   - spacer: The spacer between delimiter and string.
     ///   - ranges: The ranges where to find.
-    /// - Returns: Ranges where delimiters are, or `nil` when no delimiters was found.
-    func rangesOfBlockDelimiters(_ delimiters: Pair<String>, ranges: [NSRange]) -> [NSRange]? {
+    /// - Returns: Ranges where delimiters and spacers are, or `nil` when no delimiters was found.
+    func rangesOfBlockDelimiters(_ delimiters: Pair<String>, spacer: String, ranges: [NSRange]) -> [NSRange]? {
         
         let ranges = ranges.filter { !$0.isEmpty }
         
@@ -266,7 +273,8 @@ extension String {
         
         let beginPattern = NSRegularExpression.escapedPattern(for: delimiters.begin)
         let endPattern = NSRegularExpression.escapedPattern(for: delimiters.end)
-        let pattern = "\\A[ \t]*(\(beginPattern)).*?(\(endPattern))[ \t]*\\Z"
+        let spacerPattern = spacer.isEmpty ? "" : "(?:\(spacer))?"
+        let pattern = "\\A[ \t]*(\(beginPattern + spacerPattern)).*?(\(spacerPattern + endPattern))[ \t]*\\Z"
         let regex = try! NSRegularExpression(pattern: pattern, options: [.dotMatchesLineSeparators])
         
         let delimiterRanges = ranges